refactor: Attempt at fixing circular dependencies
