+++ src/api.ts
@@ -1,8 +1,7 @@
-export { ThemeManager } from './theme/themeManager';
-export { Theme } from './theme/theme';
+export { ThemeManager, Theme } from './theme/theme';
 export { Strategy } from './theme/strategy';
 export { Variable } from './theme/variable';
-export { ColorScheme } from './theme/colors/colorScheme';
+export { ColorScheme } from './theme/colors/color';
 export { Color, VariableColor } from './theme/colors/color';
 export {
   ColorVariant,
@@ -13,4 +12,4 @@ export {
   VariantTransformer,
   VariantsObject,
   MappedVariant,
-} from './theme/colors/variants';
+} from './theme/colors/color';
