--- src/theme/colors/colorObject.ts
@@ -1,9 +1,0 @@
-import { ColorInput } from '@ctrl/tinycolor';
-
-export interface ColorObject {
-  [name: string]: ColorInput;
-}
-
-export interface TwoLevelColorObject {
-  [name: string]: ColorInput | ColorObject;
-}
