+++ src/theme/theme.ts
@@ -1,20 +1,247 @@
+import { ColorInput } from '@ctrl/tinycolor';
 import {
-  ColorVariant,
-  OpacityVariant,
-  CustomVariant,
-  VariantType,
-  VariantTransformer,
+  TwoLevelColorObject,
   ColorScheme,
   VariableColor,
-  Variable,
-} from '../api';
-import { ColorInput } from '@ctrl/tinycolor';
-import { TwoLevelColorObject } from './colors/colorObject';
+  VariantsObject,
+  VariantInput,
+  ColorVariant,
+  OpacityVariant,
+  VariantTransformer,
+  CustomVariant,
+  Variant,
+  VariantType,
+} from './colors/color';
 import { flattenColorObject } from '../util/flattenColorObject';
-import { VariantsObject, Variant, VariantInput } from './colors/variants';
 import { isMappedVariant } from './colors/isMappedVariant';
-import { VariableInput } from './variable';
+import { VariableInput, Variable } from './variable';
+import { generateTailwindConfiguration } from '../util/generateTailwindConfiguration';
+import { generateCssConfiguration } from '../util/generateCssConfiguration';
+import { Strategy } from './strategy';
 import _ from 'lodash';
+
+/**
+ * The plugin's theme builder. It is an object that contains the
+ * configured themes for the application, as well as other settings needed
+ * to generate anything the user wants to be customized.
+ */
+export class ThemeManager {
+  private _themes: Theme[];
+  private _strategy: Strategy;
+  private _prefix: string;
+
+  /**
+   * Creates a theme manager.
+   */
+  constructor() {
+    this._themes = [];
+    this._prefix = 'theme';
+    this._strategy = Strategy.DataThemeAttribute;
+  }
+
+  /*
+  |--------------------------------------------------------------------------
+  | Default Theme Setters
+  |--------------------------------------------------------------------------
+  */
+
+  /**
+   * Defines the default theme for every color scheme.
+   */
+  setDefaultTheme(theme: Theme): this {
+    // We can't have more than one default theme.
+    if (this.getDefaultTheme()) {
+      throw new Error('Tried to set a default theme, but there was already one.');
+    }
+
+    // The default theme has no color scheme.
+    this._themes.push(theme.setDefault().setColorScheme(ColorScheme.Undefined));
+
+    return this;
+  }
+
+  setDefaultLightTheme(theme: Theme): this {
+    // We can't have more than one default light theme.
+    if (this.getDefaultLightTheme()) {
+      throw new Error('Tried to set a default light theme, but there was already one.');
+    }
+
+    this._themes.push(theme.setDefault().setColorScheme(ColorScheme.Light));
+
+    return this;
+  }
+
+  setDefaultDarkTheme(theme: Theme): this {
+    // We can't have more than one default dark theme.
+    if (this.getDefaultDarkTheme()) {
+      throw new Error('Tried to set a default dark theme, but there was already one.');
+    }
+
+    this._themes.push(theme.setDefault().setColorScheme(ColorScheme.Dark));
+
+    return this;
+  }
+
+  /*
+  |--------------------------------------------------------------------------
+  | Other Theme Adders
+  |--------------------------------------------------------------------------
+  */
+
+  /**
+   * Add a theme.
+   */
+  addTheme(theme: Theme): this {
+    this._themes.push(theme);
+
+    return this;
+  }
+
+  /**
+   * Add a light theme.
+   */
+  addLightTheme(theme: Theme): this {
+    return this.addTheme(theme.setColorScheme(ColorScheme.Light));
+  }
+
+  /**
+   * Add a dark theme.
+   */
+  addDarkTheme(theme: Theme): this {
+    return this.addTheme(theme.setColorScheme(ColorScheme.Dark));
+  }
+
+  /*
+  |--------------------------------------------------------------------------
+  | Theme getters.
+  |--------------------------------------------------------------------------
+  */
+
+  /**
+   * Gets every themes.
+   */
+  getAllThemes(): Theme[] {
+    return this._themes;
+  }
+
+  /**
+   * Get every theme, except the default one.
+   */
+  getThemes(): Theme[] {
+    return this.getThemesFor(ColorScheme.Undefined);
+  }
+
+  /**
+   * Get light-schemed themes, except the default one.
+   */
+  getLightThemes(): Theme[] {
+    return this.getThemesFor(ColorScheme.Light);
+  }
+
+  /**
+   * Get dark-schemed themes, except the default one.
+   */
+  getDarkThemes(): Theme[] {
+    return this.getThemesFor(ColorScheme.Dark);
+  }
+
+  /**
+   * Gets the default theme.
+   */
+  getDefaultTheme(): Theme | undefined {
+    return this.getDefaultThemeFor(ColorScheme.Undefined);
+  }
+
+  /**
+   * Gets the default light theme.
+   */
+  getDefaultLightTheme(): Theme | undefined {
+    return this.getDefaultThemeFor(ColorScheme.Light);
+  }
+
+  /**
+   * Gets the default dark theme.
+   */
+  getDefaultDarkTheme(): Theme | undefined {
+    return this.getDefaultThemeFor(ColorScheme.Dark);
+  }
+
+  /**
+   * Gets the default theme for the given scheme.
+   */
+  private getDefaultThemeFor(scheme: ColorScheme): Theme | undefined {
+    return this._themes.find(
+      theme => theme.isDefault() && scheme === theme.getColorScheme()
+    );
+  }
+
+  /**
+   * Gets all themes for the given scheme, except the default one.
+   */
+  private getThemesFor(scheme: ColorScheme): Theme[] {
+    return this._themes.filter(
+      theme => !theme.isDefault() && scheme === theme.getColorScheme()
+    );
+  }
+
+  /*
+  |--------------------------------------------------------------------------
+  | Strategy-related
+  |--------------------------------------------------------------------------
+  */
+
+  /**
+   * Defines the prefix used for the strategy.
+   */
+  setPrefix(prefix: string): this {
+    this._prefix = prefix;
+
+    return this;
+  }
+
+  /**
+   * Gets the prefix used for the strategy.
+   */
+  getPrefix(): string {
+    return this._prefix;
+  }
+
+  /**
+   * Defines the strategy used for theme selection.
+   */
+  setStrategy(strategy: Strategy): this {
+    this._strategy = strategy;
+
+    return this;
+  }
+
+  /**
+   * Gets the strategy used for theme selection.
+   */
+  getStrategy(): Strategy {
+    return this._strategy;
+  }
+
+  /*
+  |--------------------------------------------------------------------------
+  | Plugin
+  |--------------------------------------------------------------------------
+  */
+
+  /**
+   * Gets the Tailwind configuration for this theme manager.
+   */
+  getTailwindConfiguration(): any {
+    return generateTailwindConfiguration(this);
+  }
+
+  /**
+   * Gets an object that generates the themes' CSS inside the addBase helper.
+   */
+  getCssConfiguration(): any {
+    return generateCssConfiguration(this);
+  }
+}
 
 export class Theme {
   private _name?: string;
