+++ test/theming.test.ts
@@ -1,5 +1,3 @@
-import { ThemeManager, Theme } from '../src/api';
-import { generatePluginCss } from './generatePluginCss';
 import cssMatcher from 'jest-matcher-css';
 import _ from 'lodash';
